Added check for running SyncThing
Sometimes a user might have forgotten to remove a prior plist
file or an instance of SyncThing might already be running
for some other reason.
We are checking whether any instances are already running
and dont spawn another instance in that case.
Right now this needs to be dealt with by the user.